memset

+++ 0x07-pointers_arrays_strings/0-memset.c
@@ -13,7 +13,7 @@ char	*_memset(char *s, char b, unsigned int n)
 	int	i;
 
 	i = 0;
-	while (i < n)
+	while (i < (int) n)
 	{
 		s[i] = b;
 		i++;
